==> R/3_seurat_gsea_and_diff_expression.Rmd <==
---
title: "(3) Differential Expression & Gene Set Enrichment"
author: "Stefan Filges"
date: '`r format(Sys.Date())`'
output: 
  html_document:
    theme: sandstone
    highlight: tango
    code_folding: show
    toc: yes
    toc_depth: 3
    toc_float: yes
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Finding marker genes between clusters

Seurat can help you find markers that define clusters via differential expression. 
By default, it identifes positive and negative markers of a single cluster 
(specified in ident.1), compared to all other cells. FindAllMarkers automates 
this process for all clusters, but you can also test groups of clusters vs. 
each other, or against all cells.

The min.pct argument requires a feature to be detected at a minimum percentage 
in either of the two groups of cells, and the thresh.test argument requires a 
feature to be differentially expressed (on average) by some amount between the 
two groups. You can set both of these to 0, but with a dramatic increase in time - 
since this will test a large number of features that are unlikely to be highly 
discriminatory. As another option to speed up these computations, max.cells.per.ident 
can be set. This will downsample each identity class to have no more cells than 
whatever this is set to. While there is generally going to be a loss in power, 
the speed increases can be significiant and the most highly differentially expressed 
features will likely still rise to the top.

```{r plots_all_markers, message=FALSE}
# Packages for main analysis, clustering, data structures etc.
suppressMessages(library(Seurat, quietly = TRUE))
suppressMessages(library(mclust, quietly = TRUE))
suppressMessages(library(biomaRt, quietly = TRUE))
suppressMessages(library(msigdbr, quietly = TRUE))
suppressMessages(library(clusterProfiler, quietly = TRUE))
suppressMessages(library(destiny, quietly = TRUE))
suppressMessages(library(SummarizedExperiment, quietly = TRUE))
suppressMessages(require(clusterExperiment, quietly = TRUE))
suppressMessages(library(MAST, quietly = TRUE))
suppressMessages(library(DESeq2, quietly = TRUE))
suppressMessages(library(sctransform, quietly = TRUE))
#suppressMessages(library(bigmemory, quietly = TRUE))

# Packages for data handeling and plotting
suppressMessages(library(reshape, quietly = TRUE))
suppressMessages(library(cowplot, quietly = TRUE))
suppressMessages(library(tidyverse, quietly = TRUE))
suppressMessages(library(viridis, quietly = TRUE))
suppressMessages(library(RColorBrewer, quietly = TRUE))
suppressMessages(library(VennDiagram, quietly = TRUE))
suppressMessages(library(DT, quietly = TRUE))

# Set to TRUE if cell cycle effects were regressed out previously
filter_cell_cycle = FALSE

# define the working directory
working_directory <- '~/GitHub/SingleCellScripts/'

# set other directories based on working directory
figures_dir <- paste(working_directory, 'figures/', sep = '')
rdata_dir <- paste(working_directory, 'data/Rdata/', sep = '')

# load object
seurat_object <- paste(rdata_dir, 'seurat_object_final.Rdata', sep = '')
load(seurat_object)

# Find markers (differentially expressed genes) for each of the identity classes in a dataset
markers <- Seurat::FindAllMarkers(
  object = merged_data,
  min.pct = 0.1,
  logfc.threshold = 1
)
```

We will select the top 5 marker genes per cluster and visualise them using a 
heatmap.

```{r heatmap, fig.width=10, fig.height=8, message=FALSE}
top5 <- markers %>% group_by(cluster) %>% top_n(n = 5, wt = avg_logFC)

Seurat::DoHeatmap(
  object = merged_data,
  features = top5$gene
) + 
  Seurat::NoLegend() +
  viridis::scale_fill_viridis(
    option = 'plasma'
)

Seurat::DoHeatmap(
  object = merged_data,
  features = top5$gene, 
  group.by = "group"
) + 
  Seurat::NoLegend() +
  viridis::scale_fill_viridis(
    option = 'plasma'
)
```

## Genes upregulated in adipogenesis

The gene set for genes upregulated in adipogenesis was obtained from 
[Broad Insitute](http://software.broadinstitute.org/gsea/msigdb/cards/HALLMARK_ADIPOGENESIS.html).

The list contains 200 genes annotated as HALLMARK_ADIPOGENESIS. We visualise 
this gene set using a heatmap. Many genes were not expressed and therefore omitted
from the plot.

```{r heatmap_adipogenesis, fig.width=10, fig.height=8, message=FALSE}

database <- paste(working_directory, 'data/MSigDB/', sep = '')

adipo_markers <- read_delim(
  file = paste(database, "hallmark_genes_upregulated_during_adipogenesis.txt", sep = ''),
  delim = "\t",
  col_names = c("gene")
)

Seurat::DoHeatmap(
  object = merged_data,
  features = adipo_markers$gene,
  group.by = 'group'
) + 
  Seurat::NoLegend() +
  viridis::scale_fill_viridis(
    option = 'plasma'
)
```

## Marker genes by sample

To perform differential expression testing on a Seurat object we use the 
FindMarkers markers function and define the classes for which to perform the
comparison using:

- ident.1: Identity class to define markers for
- ident.2: A second identity class for comparison. If NULL (default) - use all other cells for comparison.

The function can use different methods to perform DE testing:

- Wilcoxon rank sum test (Seurat default)
- MAST (single-cell RNAseq specific test)
- DESeq2 with ZinB-wave weights (originally developed for bulk RNAseq data)
- Other tests such as Student's t-test, poisson and negative bionomial models

Luecken and Theis recommend to use MAST or limma for DE testing, even though
DEseq2 was also well performing for single cell data despite being orignally
developed for bulk RNAseq when using it with [ZINB-WaVE](https://www.nature.com/articles/s41467-017-02554-5) (zero inflated negative 
binomial) weights included in the model. MAST fits two-part, generalized linear models 
that are specially adapted for bimodal and/or zero-inflated single cell 
gene expression data. DE testing should be run on uncorrected data with
technical covariates included in the model.

p-value adjustment in FindMarkers is performed using bonferroni correction based on the total 
number of genes in the dataset. Other correction methods are not recommended, 
as Seurat pre-filters genes using the arguments above, reducing the number of 
tests performed. p-values should be interpreted cautiously, as the genes used for 
clustering are the same genes tested for differential expression.

When using MAST or possion/negative bionomial model it is possible to define
covariates using the latent.vars argument. We could use 'nFeature_RNA', 'nCount_RNA' 
and 'percent.mt' as latent variables and require that genes are expressed in at least
5\% of cells in at least one of the group compared.

### Genes differentially regulated between scf-WT and scf-FD
```{r scf-WT-vs-FD, fig.width=6, fig.height=4, message=FALSE, warning=FALSE}
if(filter_cell_cycle){
  markers.scfWT_vs_scfFD <- Seurat::FindMarkers(
    merged_data, 
    ident.1 = 'scf-WT', 
    ident.2 = 'scf-FD',
    test.use = 'MAST',
    min.cells.feature = 10,
    logfc.threshold = 0.1,
    latent.var = c('nFeature_RNA', 'S.Score', 'G2M.Score'),
    min.pct = 0.05,
    group.by = 'group'
  )
} else {
  markers.scfWT_vs_scfFD <- Seurat::FindMarkers(
    merged_data, 
    ident.1 = 'scf-WT', 
    ident.2 = 'scf-FD',
    test.use = 'MAST',
    min.cells.feature = 10,
    logfc.threshold = 0.1,
    latent.var = c('nFeature_RNA'),
    min.pct = 0.025,
    group.by = 'group'
  )
}

# Select and plot genes with log2-fold-change >= 1 and order genes by log2 fold change
# for plotting in the heatmap
marker_selection = markers.scfWT_vs_scfFD[abs(markers.scfWT_vs_scfFD$avg_logFC) >= 0.2, ]
marker_selection = marker_selection[order(marker_selection$avg_logFC),]

markers.scfWT_vs_scfFD <- as_tibble(
  x = markers.scfWT_vs_scfFD,
  rownames = 'genes'
)

# Output markers as interactive data table
DT::datatable(markers.scfWT_vs_scfFD, options = list(pageLength = 5))

# Plot most differentially expressed markers in heatmap
Seurat::DoHeatmap(
  object = merged_data,
  features = rownames(marker_selection),
  group.by = 'group'
  ) + 
  Seurat::NoLegend() + 
  viridis::scale_fill_viridis(
    option = 'plasma'
  )
```

### Genes differentially regulated between xen-WT and xen-FD
```{r xen-WT-vs-FD, fig.width=6, fig.height=4, message=FALSE, warning=FALSE}
if(filter_cell_cycle){
  markers.xenWT_vs_xenFD <- Seurat::FindMarkers(
    merged_data, 
    ident.1 = 'xenograft-WT', 
    ident.2 = 'xenograft-FD',
    test.use = 'MAST',
    min.cells.feature = 10,
    logfc.threshold = 0.1,
    latent.var = c('nFeature_RNA', 'S.Score', 'G2M.Score'),
    min.pct = 0.05,
    group.by = 'group'
  )
} else {
  markers.xenWT_vs_xenFD <- Seurat::FindMarkers(
    merged_data, 
    ident.1 = 'xenograft-WT', 
    ident.2 = 'xenograft-FD',
    test.use = 'MAST',
    min.cells.feature = 3,
    logfc.threshold = 0.1,
    latent.var = c('nFeature_RNA'),
    min.pct = 0.025,
    group.by = 'group'
  )
}

# Select and plot genes with log2-fold-change >= 1 and order genes by log2 fold change
# for plotting in the heatmap
marker_selection = markers.xenWT_vs_xenFD[abs(markers.xenWT_vs_xenFD$avg_logFC) >= 0.2, ]
marker_selection = marker_selection[order(marker_selection$avg_logFC),]

markers.xenWT_vs_xenFD <- as_tibble(
  x = markers.xenWT_vs_xenFD,
  rownames = 'genes'
)

# Output markers as interactive data table
DT::datatable(markers.xenWT_vs_xenFD, options = list(pageLength = 5))

# Plot most differentially expressed markers in heatmap
Seurat::DoHeatmap(
  object = merged_data,
  features = rownames(marker_selection),
  group.by = "group"
  ) + 
  Seurat::NoLegend() + 
  viridis::scale_fill_viridis(
    option = 'plasma'
  )
```
